+++ benchmarks/eval_benchmarks/cve_bench/__init__.py
@@ -1,0 +1,3 @@
+from .cve_bench import cve_bench
+
+__all__ = ["cve_bench"]